move segmenter size estimator to bitmapsegmenter
